+++ src/presenters/pages/login/index.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const LoginPage = () => {
+  return <div className="content">For new work</div>;
+};
+
+export default LoginPage;
